fix: fleet oracle/gate no longer hang when cargo holds the build lock
run_bounded's timeout branch called child.wait() after child.kill(). When
cargo (the gate subprocess) is blocked on flock() waiting for the cargo
build lock held by the outer test runner, SIGKILL is queued but not
delivered until flock() wakes -- putting child.wait() in D-state for up
to 33 seconds. Fleet oracle then exceeded the 20-second test deadline.

Fix: spawn a detached background thread to reap the zombie instead of
blocking the call site. The SIGKILL is still sent immediately; the zombie
is cleaned up asynchronously and doesn't block the next gate or exit.

Also fix: CoverageProvider::coverage_percent() now returns u64 (was f64),
consistent with the no-float rule (P2 META-L8-REVIEW finding). Test
literals changed from 100.0/70.0 to 100/70.

Verified: hangs_terminate_real_binary 3/3 pass in 3.76s (was 53s).

+++ fleet/crates/verify/src/tests.rs
@@ -11,7 +11,7 @@ fn candidate(gates: Vec<GateSpec>) -> ReviewedCandidate {
     }
 }
 fn spec(id: &str) -> GateSpec { GateSpec { id: id.into(), command: "true".into(), args: vec![] } }
-fn run(c: &ReviewedCandidate, pass: bool, cov: f64, fp: FakeFindingsProvider)
+fn run(c: &ReviewedCandidate, pass: bool, cov: u64, fp: FakeFindingsProvider)
     -> crate::GateEvidence
 {
     let runner = if pass { FakeGateRunner::passing() } else { FakeGateRunner::failing() };
@@ -20,7 +20,7 @@ fn run(c: &ReviewedCandidate, pass: bool, cov: f64, fp: FakeFindingsProvider)
 
 #[test]
 fn reviewed_candidate_runs_deterministic_gates() {
-    let ev = run(&candidate(vec![spec("g1")]), true, 100.0, FakeFindingsProvider::empty());
+    let ev = run(&candidate(vec![spec("g1")]), true, 100, FakeFindingsProvider::empty());
     assert!(!ev.gate_results.is_empty());
     assert!(ev.checked > 0);
     assert_eq!(ev.checked, ev.total);
@@ -29,7 +29,7 @@ fn reviewed_candidate_runs_deterministic_gates() {
 
 #[test]
 fn failing_gate_produces_gate_evidence_not_panic() {
-    let ev = run(&candidate(vec![spec("g1")]), false, 100.0, FakeFindingsProvider::empty());
+    let ev = run(&candidate(vec![spec("g1")]), false, 100, FakeFindingsProvider::empty());
     assert!(!ev.passed);
     assert!(!ev.failures.is_empty());
 }
@@ -38,7 +38,7 @@ fn failing_gate_produces_gate_evidence_not_panic() {
 fn coverage_below_floor_marks_gate_failed() {
     let mut c = candidate(vec![spec("g1")]);
     c.coverage_floor = Some(80);
-    let ev = run(&c, true, 70.0, FakeFindingsProvider::empty());
+    let ev = run(&c, true, 70, FakeFindingsProvider::empty());
     let cov = ev.gate_results.iter().find(|r| r.id == "coverage").unwrap();
     assert!(!cov.passed);
     let msg = cov.failure_message.as_deref().unwrap_or("");
@@ -48,7 +48,7 @@ fn coverage_below_floor_marks_gate_failed() {
 
 #[test]
 fn secret_finding_blocks_gate() {
-    let ev = run(&candidate(vec![spec("g1")]), true, 100.0,
+    let ev = run(&candidate(vec![spec("g1")]), true, 100,
         FakeFindingsProvider::with_critical("secrets.txt"));
     assert!(!ev.passed);
     assert!(!ev.findings.is_empty());
@@ -58,7 +58,7 @@ fn secret_finding_blocks_gate() {
 fn evidence_digest_mismatch_refused() {
     let mut c = candidate(vec![spec("g1")]);
     c.output_digest = Some("wrong-digest".into());
-    let ev = run(&c, true, 100.0, FakeFindingsProvider::empty());
+    let ev = run(&c, true, 100, FakeFindingsProvider::empty());
     assert!(!ev.passed);
     let found = ev.failures.iter().any(|f| f.contains("mismatch") || f.contains("digest"));
     assert!(found, "failures: {:?}", ev.failures);
